requirements.txt
aiocache==0.12.3
aiofiles==24.1.0
aioredis==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==5.0.1
bcrypt==4.0.1
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==43.0.3
distlib==0.3.9
dnspython==2.7.0
dulwich==0.21.7
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cache==0.1.0
fastapi-cache2==0.1.3.4
fastapi-cli==0.0.5
fastjsonschema==2.20.0
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
keyring==24.3.1
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
packaging==24.2
passlib==1.7.4
pexpect==4.9.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry-core==1.9.1
poetry-plugin-export==1.8.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.6.1
pydantic_core==2.20.1
Pygments==2.18.0
PyMySQL==1.1.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.10.1
redis==5.0.8
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqladmin==0.18.0
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.37.2
tomli==2.1.0
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typer==0.13.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.1
watchfiles==0.24.0
websockets==14.0
WTForms==3.1.2
zipp==3.21.0